+++ tests/crc_test.c
@@ -73,7 +73,7 @@ static int s_test_known_crc_32(
     ASSERT_HEX_EQUALS(expected_crc, result, "%s(%s)", func_name, data_name);
 
     // Compute the residue of the buffer (the CRC of the buffer plus its CRC) - will always be a constant value
-    uint32_t residue = func((const uint8_t *)&result, 4, result); // assuming little endian
+    uint32_t residue = (uint32_t)func((const uint8_t *)&result, 4, result); // assuming little endian
     ASSERT_HEX_EQUALS(expected_residue, residue, "len %d residue %s(%s)", length, func_name, data_name);
 
     // chain the crc computation so 2 calls each operate on about 1/2 of the buffer
